implemented parts of the nvs API

+++ FLASH_SPI.h
@@ -18,7 +18,7 @@ uint8_t FLASH_ERASE_PARTITION(const char* partition_name);
 
 PARTITION_INFO_t FLASH_GET_PARTITION_INFO(const char* partition_name);
 
-uint8_t FLASH_DOES_KEY_EXIST(const char* partition_name, const char* namespace, const char* blob_name);
+size_t FLASH_DOES_KEY_EXIST(const char* partition_name, const char* namespace, const char* blob_name);
 
 uint8_t FLASH_WRITE_TO_PARTITION(const char* partition_name, const char* namespace, const char* blob_name, void* data, unsigned long size);
